mudanças no comportamento da sidebar e iniciando listagem de produtos
