+++ frontend/src/pages/Home.tsx
@@ -1,21 +1,20 @@
 import { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
+import { useCookies } from 'react-cookie';
+
 import IconServices from '../public/assets/home-page/icons/services_icon.svg';
 import IconProducts from '../public/assets/home-page/icons/products_icon.svg';
 import IconUsers from '../public/assets/home-page/icons/user_icon.svg';
 import IconClients from '../public/assets/home-page/icons/clients_icon.svg';
-import IconMenu from '../public/assets/home-page/icons/menu_icon.svg';
-import InfoCard from '../components/infocard'; 
+import InfoCard from '../components/infocard';
 import Sidebar from '../components/sidebar';
 import TopNav from '../components/top_nav';
-import { useCookies } from 'react-cookie';
 
 export const Home = () => {
 	const [cookies, setCookie] = useCookies(['id', 'jwt']);
 	console.log(cookies.id);
 	console.log(cookies);
 
-	const [sidebarOpen, setSidebarOpen] = useState(false);
 	const [statistics, setStatistics] = useState({
 		openOrders: 0,
 		clients: 0,
@@ -24,10 +23,6 @@ export const Home = () => {
 	});
 
 	const today = new Date().toLocaleDateString('pt-BR');
-
-	const toggleSidebar = () => {
-		setSidebarOpen(!sidebarOpen);
-	};
 
 	useEffect(() => {
 		const fetchData = async () => {
@@ -41,21 +36,7 @@ export const Home = () => {
 
 	return (
 		<div className="flex h-screen">
-			<Sidebar sidebarOpen={sidebarOpen} toggleSidebar={toggleSidebar} />
-			{!sidebarOpen && (
-				<button
-					className="lg:hidden fixed top-10 left-5 z-40 p-2 bg-white rounded-full"
-					onClick={toggleSidebar}
-				>
-					<img className="w-6" src={IconMenu} alt="icone do menu" />
-				</button>
-			)}
-			{sidebarOpen && (
-				<div
-					className="fixed inset-0 bg-black opacity-50 lg:hidden"
-					onClick={toggleSidebar}
-				></div>
-			)}
+			<Sidebar />
 			<main className="flex-1 p-10 bg-blue-200 space-y-10">
 				<header className="flex justify-between">
 					<div className="pt-16 md:pt-16 lg:pt-0">
